extract methods for loader as QueueingService
