+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/source/OnSourceMenuSelected.scala
@@ -4,14 +4,12 @@ import android.support.v4.app.FragmentActivity
 import android.support.v7.app.AlertDialog
 import x7c1.linen.glue.res.layout.{SettingSourceAttach, SettingSourceAttachRowItem}
 import x7c1.linen.glue.service.ServiceControl
-import x7c1.linen.glue.service.ServiceLabel.Updater
 import x7c1.linen.modern.display.settings.SourceMenuSelected
 import x7c1.linen.modern.init.settings.preset.AttachSourceDialog
-import x7c1.linen.scene.updater.UpdaterMethods
+import x7c1.linen.scene.loader.crawling.QueueingService
 import x7c1.wheat.ancient.context.ContextualFactory
 import x7c1.wheat.ancient.resource.ViewHolderProviderFactory
 import x7c1.wheat.macros.fragment.FragmentFactory
-import x7c1.wheat.macros.intent.ServiceCaller
 import x7c1.wheat.modern.menu.popup.{PopupMenuBox, PopupMenuItem}
 
 
@@ -27,10 +25,7 @@ class OnSourceMenuSelected(
       createAttachSourceDialog(event) showIn activity
     }
     def toLoad = PopupMenuItem("Load now"){ _ =>
-      ServiceCaller.using[UpdaterMethods].
-        startService(activity, activity getClassOf Updater){
-          _ loadSource event.selectedSourceId
-        }
+      QueueingService(activity).loadSource(event.selectedSourceId)
     }
     val items = Seq(
       toAttach,
