+++ linen-scene/src/main/scala/x7c1/linen/scene/channel/menu/OnChannelMenuSelected.scala
@@ -3,11 +3,10 @@ package x7c1.linen.scene.channel.menu
 import android.app.Activity
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.glue.service.ServiceControl
-import x7c1.linen.glue.service.ServiceLabel.Updater
 import x7c1.linen.repository.channel.my.MyChannel
 import x7c1.linen.repository.channel.preset.SettingPresetChannel
-import x7c1.linen.scene.updater.UpdaterMethods
-import x7c1.wheat.macros.intent.{LocalBroadcaster, ServiceCaller}
+import x7c1.linen.scene.loader.crawling.QueueingService
+import x7c1.wheat.macros.intent.LocalBroadcaster
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.formatter.ThrowableFormatter
 import x7c1.wheat.modern.menu.popup.PopupMenuItem
@@ -50,10 +49,7 @@ class MenuItemFactory(
   def toLoadSources(channelId: Long): PopupMenuItem =
     PopupMenuItem("Load all sources"){ _ =>
       Log info s"[init] channel:$channelId"
-      ServiceCaller.using[UpdaterMethods].
-        startService(activity, activity getClassOf Updater){
-          _ loadChannelSources (channelId, accountId)
-        }
+      QueueingService(activity).loadChannelSources(channelId, accountId)
     }
 
   def toDeleteChannel(
