+++ linen-scene/src/main/scala/x7c1/linen/scene/loader/crawling/SubscribedChannelsLoader.scala
@@ -3,23 +3,19 @@ package x7c1.linen.scene.loader.crawling
 import android.content.Context
 import x7c1.linen.database.control.DatabaseHelper
 import x7c1.linen.database.struct.AccountIdentifiable
-import x7c1.linen.glue.service.{ServiceControl, ServiceLabel}
+import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.repository.channel.subscribe.SubscribedChannel
-import x7c1.linen.scene.updater.UpdaterMethods
-import x7c1.wheat.macros.intent.ServiceCaller
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.formatter.ThrowableFormatter.format
 
 class SubscribedChannelsLoader private (
-  context: Context,
-  control: ServiceControl,
+  context: Context with ServiceControl,
   helper: DatabaseHelper ){
 
   def execute[A: AccountIdentifiable](account: A): Unit = {
     Log info s"[init]"
 
     val accountId = implicitly[AccountIdentifiable[A]] toId account
-    val caller = ServiceCaller.using[UpdaterMethods]
 
     helper.selectorOf[SubscribedChannel] traverseOn account match {
       case Left(e) => Log error format(e){"[failed]"}
@@ -27,9 +23,7 @@ class SubscribedChannelsLoader private (
         sequence.toSeq foreach { channel =>
           Log info s"$channel"
 
-          caller.startService(context, control getClassOf ServiceLabel.Updater){
-            _.loadChannelSources(channel.channelId, accountId)
-          }
+          QueueingService(context).loadChannelSources(channel.channelId, accountId)
         }
         sequence.closeCursor()
     }
@@ -41,6 +35,6 @@ object SubscribedChannelsLoader {
     context: Context with ServiceControl,
     helper: DatabaseHelper): SubscribedChannelsLoader = {
 
-    new SubscribedChannelsLoader(context, context, helper)
+    new SubscribedChannelsLoader(context, helper)
   }
 }
